+++ peppol-common/src/main/java/no/difi/vefa/peppol/common/api/Icd.java
@@ -1,0 +1,13 @@
+package no.difi.vefa.peppol.common.api;
+
+import no.difi.vefa.peppol.common.model.Scheme;
+
+public interface Icd {
+    String getIdentifier();
+
+    String getIcd();
+
+    Scheme getScheme();
+
+    boolean isDeprecated();
+}