+++ peppol-common/src/main/java/no/difi/vefa/peppol/common/code/Mode.java
@@ -1,4 +1,4 @@
-package no.difi.vefa.peppol.common.api;
+package no.difi.vefa.peppol.common.code;
 
 public enum Mode {
     PRODUCTION,