+++ peppol-lookup/src/test/java/no/difi/vefa/peppol/lookup/reader/BdxrReaderTest.java
@@ -32,7 +32,7 @@ public class BdxrReaderTest {
     public void serviceMetadata() throws Exception {
         ServiceMetadata result = reader.parseServiceMetadata(new FetcherResponse(getClass().getResourceAsStream("/bdxr-servicemetadata-9908-810418052.xml"), null));
 
-        ProcessIdentifier processIdentifier = new ProcessIdentifier("urn:www.cenbii.eu:profile:bii04:ver1.0", "cenbii-procid-ubl");
+        ProcessIdentifier processIdentifier = new ProcessIdentifier("urn:www.cenbii.eu:profile:bii04:ver1.0");
 
         try {
             result.getEndpoint(processIdentifier, TransportProfile.START);