+++ peppol-lookup/src/test/java/no/difi/vefa/peppol/lookup/reader/BusdoxReaderTest.java
@@ -32,7 +32,7 @@ public class BusdoxReaderTest {
     public void serviceMetadata() throws Exception {
         ServiceMetadata result = reader.parseServiceMetadata(new FetcherResponse(getClass().getResourceAsStream("/busdox-servicemetadata-9908-991825827.xml"), null));
 
-        ProcessIdentifier processIdentifier = new ProcessIdentifier("urn:www.cenbii.eu:profile:bii05:ver2.0", "cenbii-procid-ubl");
+        ProcessIdentifier processIdentifier = new ProcessIdentifier("urn:www.cenbii.eu:profile:bii05:ver2.0");
 
         try {
             result.getEndpoint(processIdentifier, TransportProfile.START);