+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Adding Mode enum.
 * Adding no.difi.vefa.peppol.common.model.Scheme and related changes.
 * Changing namespace of exceptions.
+* Adding Icd interface and ISO6523 enum, rewriting ParticipantIdentifier to use it.
 
 ## 0.9.3